pyproject.toml
[project]
name = "research-material-science"
version = "0.1.0"
description = "Material science research: perovskite stability descriptors"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "openai>=1.30.0",
    "hydra-core>=1.3.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "graphviz>=0.21",
    "pysr>=1.5.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pre-commit>=3.7.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reproduce_evidence"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["legacy/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]
